+++ source/split_mkv/plugin.py
@@ -758,8 +758,9 @@ def get_chapters_from_credits(srcpath, duration, tmp_dir, settings):
             except IndexError:
                 firstfile = ''
             try:
-                lastfile = [i for i in range(len(density)-3) if i > lastfile and density[i] == 0 and density[i-1] > 10 and density[i-2] > 25 and density[i-3] > 40][0]
-            except IndexError:
+                lastfile = [i for i in range(len(density)-3) if (firstfile != '' and i > firstfile and density[i] == 0 and density[i-1] > 10 and density[i-2] > 25 and density[i-3] > 40) or 
+                            (if firstfile =='' and density[i] == 0 and density[i-1] > 10 and density[i-2] > 25 and density[i-3] > 40)][0]
+            except (IndexError, TypeError):
                 lastfile = ''
             try:
                 falseend = [i for i in range(len(density)-3) if i> lastfile and density[i] == 0 and density[i+1] > 30 and density[i+2] > 100 and density[i+3] > 100][0]